Add in the MarkdownReader class.
Mostly a giant wad of code to allow the current rST templates to play nice with
markdown files. Hopefully, this Markdown class not only reads valid Markdown
and valid Markdown metadata, but converts the limited subset of rST metadata
into valid Markdown metadata too. This is achieved by a combination of regex
parsing and filtering of lines, along the way bailing if it looks like it's
valid Markdown

Note that I've commented out the dependency on python-markdown in global.txt;
this is because I see this as a feature request to a piece of software and
figure that there may be a design decision behind this. I've also done my best
to make this code run _without_ the need for the Markdown renderer, however
obviously I do feel like it'd be best for this to be shipped as a dependency!

Also I like my regex string splitter better than I like the original
"string.split(',')"; credited from stackoverflow. Feel free to use this in
place of your splitter, which won't play nice with things I intentionally want
a comma in!

Signed-off-by: Ben Stott <bgbnbigben@contextualsystems.com>