Fix gkmngrgn/radpress#12.
Creates a new templatetag that takes the current context (which by virtue of
TEMPLATE_CONTEXT_PROCESSORS contains the current request object (plus all the
defaults -- I think this is a stupid system but I don't know that django allows
me to use += ...)) and builds the correct absolute URI for the given path. This
is then placed straight into the template and not attempted to be
parsed/corrected by the javascript code.

Signed-off-by: Ben Stott <bgbnbigben@contextualsystems.com>